Migrate spatial4j-core to be the new top-level sole module.  And upgrade randomizedtesting dependency version.
