+++ src/main/java/com/spatial4j/core/context/SpatialContext.java
