+++ src/main/java/com/spatial4j/core/context/jts/JtsSpatialContext.java
