+++ src/main/java/com/spatial4j/core/context/jts/JtsSpatialContextFactory.java
