+++ src/main/java/com/spatial4j/core/distance/DistanceCalculator.java
