+++ src/main/java/com/spatial4j/core/distance/package-info.java
