+++ src/main/java/com/spatial4j/core/exception/InvalidShapeException.java
