+++ src/main/java/com/spatial4j/core/io/LineReader.java
@@ -17,13 +17,7 @@
 
 package com.spatial4j.core.io;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.Reader;
+import java.io.*;
 import java.util.Iterator;
 
 public abstract class LineReader<T> implements Iterator<T> {
