+++ src/main/java/com/spatial4j/core/io/geonames/Geoname.java
