+++ src/main/java/com/spatial4j/core/io/sample/SampleData.java
