+++ src/main/java/com/spatial4j/core/io/sample/SampleDataWriter.java
@@ -17,19 +17,14 @@
 
 package com.spatial4j.core.io.sample;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStreamWriter;
-import java.io.PrintWriter;
-
 import com.spatial4j.core.context.SpatialContext;
 import com.spatial4j.core.context.jts.JtsSpatialContext;
 import com.spatial4j.core.shape.Shape;
-
 import com.vividsolutions.jts.geom.Envelope;
 import com.vividsolutions.jts.geom.Geometry;
 import com.vividsolutions.jts.simplify.TopologyPreservingSimplifier;
+
+import java.io.*;
 
 public class SampleDataWriter {
