+++ src/main/java/com/spatial4j/core/package-info.java
