+++ src/main/java/com/spatial4j/core/shape/Circle.java
