+++ src/main/java/com/spatial4j/core/shape/Point.java
