+++ src/main/java/com/spatial4j/core/shape/Rectangle.java
