+++ src/main/java/com/spatial4j/core/shape/SpatialRelation.java
