+++ src/main/java/com/spatial4j/core/shape/impl/PointImpl.java
