+++ src/main/java/com/spatial4j/core/shape/jts/JtsPoint.java
