+++ src/main/java/com/spatial4j/core/util/GeohashUtils.java
