+++ src/test/java/com/spatial4j/core/context/SpatialContextFactoryTest.java
