+++ src/test/java/com/spatial4j/core/distance/CompareRadiansSnippet.java
