+++ src/test/java/com/spatial4j/core/distance/TestDistances.java
