+++ src/test/java/com/spatial4j/core/io/ShapeReadWriterTest.java
