+++ src/test/java/com/spatial4j/core/shape/RoundingDistCalc.java
