+++ src/test/java/com/spatial4j/core/util/TestGeohashUtils.java
