+++ src/test/resources/fiji.wkt.txt
