+++ src/test/resources/russia.wkt.txt
